refactor: Naming 오타 수정
